pyproject.toml
[project]
name = "mcp-client-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.75.0",
    "fastapi>=0.104.0",
    "mcp>=1.24.0",
    "openai>=2.11.0",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
]
